Rename dnf5 config option getters
The option getters were renamed in dnf5 in commit:
rpm-software-management/dnf5/commit/740fd279d283156b8e9c3525a16142f2fffda291

There is also another option and that is to change the DNFConfigWrapper.
However, this class is described as a temporary wrapper, so further
investigation will be needed around this. For now, renaming the options
seems sufficient.